[CLEAN] project_accounting: externalisation des données d'initialisation historiques 2026
- Déplacement de la fonction data_init_2026 et de ses dictionnaires volumineux vers models/old_data_init_2026.py.
- Suppression de la fonction dans project.progress pour simplifier le modèle.
- Conservation des données pour archive et audit historique uniquement.
