+++ project_accounting/models/old_data_init_2026.py
@@ -1,0 +1,727 @@
+import logging
+import datetime
+
+_logger = logging.getLogger(__name__)
+
+# Ce fichier contient l'ancienne fonction d'initialisation des données 2026.
+# Elle a été déplacée ici pour l'historique et n'est plus appelée par le modèle project.progress.
+
+    def data_init_2026(self):
+        cumuls_31_12_2025 = {
+'99TS005' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 30609.92, 'ca_autre' : 0},
+'99TS004' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'99TS003' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 21422.22, 'ca_autre' : 0},
+'99TS002' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 37245, 'ca_autre' : 0},
+'99TS001' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 12164.76, 'ca_autre' : 0},
+'26TS085' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS084' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS083' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS082' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS081' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS080' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS079' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS078' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS077' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS076' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS075' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS074' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS073' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS072' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS071' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS070' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS069' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS068' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS067' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS066' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS065' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS064' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS063' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS062' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS061' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS060' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS059' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS058' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS057' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS056' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS055' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS054' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS053' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS052' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS051' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS050' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS049' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS048' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS047' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS046' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS045' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS044' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS043' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS042' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS041' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS038' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS037' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS031' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS029' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS028' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS027' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS026' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS017' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS016' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS015' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS014' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS013' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS012' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS011' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS010' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS009' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS008' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS007' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS002' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS001' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS379' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS376' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS374' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS373' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS372' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS371' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS370' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS369' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS368' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS367' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS365' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS364' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS363' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS362' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS361' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS359' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS358' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS357' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS356' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS355' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS354' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS353' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS352' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS351' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS350' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS345' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS344' : {'internal' : 16172.25, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS343' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS340' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS339' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS338' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS333' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS332' : {'internal' : 6400, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS331' : {'internal' : 11084.43, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS330' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS329' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS328' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS327' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS326' : {'internal' : 4519.88, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 33.58},
+'25TS325' : {'internal' : 14486.04, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS323' : {'internal' : 2005.25, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS322' : {'internal' : 7.27595761418343E-12, 'ogures' : 95606.65, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 50553.35, 'prestation' : 0, 'ca_autre' : 0},
+'25TS321' : {'internal' : 9414.01, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS320' : {'internal' : 1473.65, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS319' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS318' : {'internal' : 15396, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS317' : {'internal' : 34670, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS316' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS315' : {'internal' : 7540.66, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 534.34},
+'25TS311' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS310' : {'internal' : 85000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS309' : {'internal' : 4985, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS308' : {'internal' : 7949.4, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 1380, 'prestation' : 0, 'ca_autre' : 70.6},
+'25TS307' : {'internal' : 25910, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 95},
+'25TS306' : {'internal' : 120309.72, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 2600, 'prestation' : 0, 'ca_autre' : 341},
+'25TS303' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS300' : {'internal' : 1561.7, 'ogures' : 36908, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 245.3},
+'25TS299' : {'internal' : 50000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS298' : {'internal' : 4000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS297' : {'internal' : 16000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS295' : {'internal' : 26000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS294' : {'internal' : 62130, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS293' : {'internal' : 8091.88, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 102.72},
+'25TS292' : {'internal' : 53982.31, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 1976.59},
+'25TS291' : {'internal' : -2.91038304567337E-11, 'ogures' : 34311.8, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 185829.6, 'prestation' : 0, 'ca_autre' : 0},
+'25TS290' : {'internal' : 21156.5, 'ogures' : 23000, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS288' : {'internal' : 19421.04, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 28.96},
+'25TS287' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS285' : {'internal' : 46300, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS284' : {'internal' : 56414.23, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 1865.77},
+'25TS283' : {'internal' : 33000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS282' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS281' : {'internal' : 22086, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 519.48},
+'25TS280' : {'internal' : 48100, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS277' : {'internal' : 29529.11, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 51980, 'prestation' : 0, 'ca_autre' : 48.89},
+'25TS276' : {'internal' : 125500, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 21000, 'prestation' : 0, 'ca_autre' : 0},
+'25TS275' : {'internal' : 20609.62, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 390.38},
+'25TS274' : {'internal' : 106090, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 28670, 'prestation' : 0, 'ca_autre' : 0},
+'25TS273' : {'internal' : 33500, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS272' : {'internal' : 34166.2, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS271' : {'internal' : 44770, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS270' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS269' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS268' : {'internal' : 21750, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS267' : {'internal' : 694.01, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS266' : {'internal' : 4286.3, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 69.45},
+'25TS264' : {'internal' : 249336.31, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 373.69},
+'25TS263' : {'internal' : 16000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS262' : {'internal' : 103345, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS261' : {'internal' : 63083.06, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 61.94},
+'25TS257' : {'internal' : 37000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS256' : {'internal' : 16387.8, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS255' : {'internal' : 22960, 'ogures' : 94500, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS254' : {'internal' : 21875, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS253' : {'internal' : 38894.5, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 25100, 'prestation' : 0, 'ca_autre' : 105.5},
+'25TS252' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 4500, 'prestation' : 0, 'ca_autre' : 0},
+'25TS251' : {'internal' : 91323.3, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 76.7},
+'25TS250' : {'internal' : 11520, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS249' : {'internal' : 33631.07, 'ogures' : 126738.93, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 125400, 'prestation' : 0, 'ca_autre' : 0},
+'25TS248' : {'internal' : 167388.5, 'ogures' : 108766.5, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 170400, 'prestation' : 0, 'ca_autre' : 0},
+'25TS247' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 195120, 'prestation' : 0, 'ca_autre' : 0},
+'25TS244' : {'internal' : 23996.06, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 533.94},
+'25TS242' : {'internal' : 231828, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 88850, 'prestation' : 0, 'ca_autre' : 896.05},
+'25TS241' : {'internal' : 20267.5, 'ogures' : 44682.5, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS240' : {'internal' : 26692.98, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 3852.48},
+'25TS238' : {'internal' : 69266, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS228' : {'internal' : 199955.06, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 44.94},
+'25TS225' : {'internal' : 39800, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS219' : {'internal' : 106000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS218' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 245000, 'prestation' : 0, 'ca_autre' : 0},
+'25TS216' : {'internal' : 105289.68, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS206' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS204' : {'internal' : 108852.18, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 17209.16, 'prestation' : 0, 'ca_autre' : 1188.66},
+'25TS203' : {'internal' : 128250, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 178.45},
+'25TS197' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS196' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS195' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS174' : {'internal' : 27186.14, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 383.86},
+'25TS173' : {'internal' : 0, 'ogures' : 132000, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS170' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS161' : {'internal' : 2250, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 2250, 'prestation' : 0, 'ca_autre' : 0},
+'25TS151' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS148' : {'internal' : 0, 'ogures' : 150000, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS142' : {'internal' : 7696.52, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 6163.48, 'prestation' : 0, 'ca_autre' : 0},
+'25TS141' : {'internal' : 133848, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS137' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS136' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS134' : {'internal' : -489.54, 'ogures' : 393602.87, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 157682.13, 'prestation' : 0, 'ca_autre' : 489.54},
+'25TS131' : {'internal' : 60390, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 759.96},
+'25TS130' : {'internal' : 48770.5, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 9.5},
+'25TS126' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS121' : {'internal' : 56473.44, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 1000, 'prestation' : 0, 'ca_autre' : 276.56},
+'25TS120' : {'internal' : 147706.93, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 4601.98},
+'25TS115' : {'internal' : 97267.2, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 240.2},
+'25TS114' : {'internal' : 5227.5, 'ogures' : 41535, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS095' : {'internal' : 90326.8, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 2014.33},
+'25TS091' : {'internal' : 250375.5, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 124.5},
+'25TS082' : {'internal' : 25000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS079' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 12000, 'prestation' : 0, 'ca_autre' : 0},
+'25TS016' : {'internal' : 83947.24, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 127.09},
+'25TS002' : {'internal' : 56500, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'24TS377' : {'internal' : 91410, 'ogures' : 0, 'mc2i' : 28000, 'inditto' : 0, 'autre_ST' : 30590, 'prestation' : 0, 'ca_autre' : 0},
+'24TS359' : {'internal' : 100000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'24TS357' : {'internal' : 259026.42, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 4090.58},
+'24TS184' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'24TS023' : {'internal' : 84645.05, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 13000, 'prestation' : 0, 'ca_autre' : 2850.95},
+'23294' : {'internal' : 13900, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+            }
+
+        ca_periode_01_2026 = {
+'99TS005' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 1884.94, 'ca_autre' : 0},
+'99TS004' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'99TS003' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : -1500, 'ca_autre' : 0},
+'99TS002' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'99TS001' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 73.4, 'ca_autre' : 0},
+'26TS085' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS084' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS083' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS082' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS081' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS080' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS079' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS078' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS077' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS076' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS075' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS074' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS073' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS072' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS071' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS070' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS069' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS068' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS067' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS066' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS065' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS064' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS063' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS062' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS061' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS060' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS059' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS058' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS057' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS056' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS055' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS054' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS053' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS052' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS051' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS050' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS049' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS048' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS047' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS046' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS045' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS044' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS043' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS042' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS041' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS038' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS037' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS031' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS029' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS028' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS027' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS026' : {'internal' : 1764.7, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS017' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS016' : {'internal' : 7000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS015' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS014' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS013' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS012' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS011' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS010' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS009' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS008' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 3600, 'prestation' : 0, 'ca_autre' : 0},
+'26TS007' : {'internal' : 2000, 'ogures' : 17800, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS002' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 11400, 'prestation' : 0, 'ca_autre' : 0},
+'26TS001' : {'internal' : 0, 'ogures' : 20400, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS379' : {'internal' : 19409.09, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS376' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS374' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS373' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS372' : {'internal' : 11722.97, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS371' : {'internal' : 14555, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS370' : {'internal' : 9553.57, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS369' : {'internal' : 40000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS368' : {'internal' : 6900, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS367' : {'internal' : 18860.09, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 7475, 'prestation' : 0, 'ca_autre' : 564.91},
+'25TS365' : {'internal' : 10000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS364' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS363' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS362' : {'internal' : 3162.5, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS361' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS359' : {'internal' : 36877.19, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS358' : {'internal' : 44889.83, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS357' : {'internal' : 33000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS356' : {'internal' : 18000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS355' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS354' : {'internal' : 32192.47, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 807.53},
+'25TS353' : {'internal' : 6310.34, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS352' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS351' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS350' : {'internal' : 15684.93, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS345' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS344' : {'internal' : 23957.25, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS343' : {'internal' : 12375, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS340' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS339' : {'internal' : 37500, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS338' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS333' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS332' : {'internal' : 14100, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS331' : {'internal' : 9000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS330' : {'internal' : 0, 'ogures' : 31355, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS329' : {'internal' : 22266.02, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 33.98},
+'25TS328' : {'internal' : -1.02318153949454E-12, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 9223.38, 'prestation' : 0, 'ca_autre' : 176.62},
+'25TS327' : {'internal' : 68500, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS326' : {'internal' : 24525.7, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 229.25},
+'25TS325' : {'internal' : 6480.88, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS323' : {'internal' : 50887, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 388},
+'25TS322' : {'internal' : -2.27373675443232E-13, 'ogures' : -2747.22, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : -1277.78, 'prestation' : 0, 'ca_autre' : 0},
+'25TS321' : {'internal' : 46300, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS320' : {'internal' : 3710.35, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS319' : {'internal' : 15200, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS318' : {'internal' : 5000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS317' : {'internal' : 11037.5, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS316' : {'internal' : 7.105427357601E-14, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 8948.83, 'prestation' : 0, 'ca_autre' : 51.17},
+'25TS315' : {'internal' : 3592.84, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 707.16},
+'25TS311' : {'internal' : 9212.9, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS310' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS309' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS308' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS307' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS306' : {'internal' : 61429, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 3571, 'prestation' : 0, 'ca_autre' : 0},
+'25TS303' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS300' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS299' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS298' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS297' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS295' : {'internal' : 10600, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS294' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS293' : {'internal' : 16750, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS292' : {'internal' : 10000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS291' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS290' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS288' : {'internal' : -28.96, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 28.96},
+'25TS287' : {'internal' : 24600, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS285' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS284' : {'internal' : 312.5, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS283' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS282' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS281' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS280' : {'internal' : 13900, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS277' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS276' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS275' : {'internal' : 3703.12, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS274' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS273' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS272' : {'internal' : 6500, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS271' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS270' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS269' : {'internal' : 12250, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS268' : {'internal' : 24900, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS267' : {'internal' : 11400, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS266' : {'internal' : 14000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS264' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS263' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS262' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS261' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS257' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS256' : {'internal' : 4000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS255' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS254' : {'internal' : 10097, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 103},
+'25TS253' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS252' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS251' : {'internal' : -16960, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS250' : {'internal' : 2300, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS249' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS248' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS247' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS244' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS242' : {'internal' : -13050, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 13050, 'prestation' : 0, 'ca_autre' : 0},
+'25TS241' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS240' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS238' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS228' : {'internal' : 12160, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS225' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS219' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS218' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS216' : {'internal' : 20000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS206' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS204' : {'internal' : 59146.74, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 1653.26},
+'25TS203' : {'internal' : -3375, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS197' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS196' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS195' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS174' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS173' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS170' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS161' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS151' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS148' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS142' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS141' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS137' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS136' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS134' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS131' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS130' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS126' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS121' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS120' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS115' : {'internal' : 192.6, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS114' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS095' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS091' : {'internal' : 21800, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS082' : {'internal' : 2200, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS079' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS016' : {'internal' : 813.56, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS002' : {'internal' : 2000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'24TS377' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'24TS359' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'24TS357' : {'internal' : 26000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'24TS184' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'24TS023' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'23294' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+    }
+
+        ca_periode_02_2026 = {
+'99TS005' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 1884.94, 'ca_autre' : 0},
+'99TS004' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'99TS003' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'99TS002' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 4500, 'ca_autre' : 0},
+'99TS001' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 1043.82, 'ca_autre' : 0},
+'26TS106' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS105' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS104' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS103' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS102' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS101' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS100' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS099' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS098' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS097' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS096' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS094' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS093' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS092' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS091' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS090' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS089' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS088' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS087' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS085' : {'internal' : 400, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS084' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS083' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS082' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS081' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS080' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS079' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS078' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS075' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS074' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS073' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS072' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS071' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS070' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS069' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS068' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS067' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS066' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS065' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS064' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS063' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS062' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS061' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS060' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS059' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS058' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS057' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS056' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS055' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS054' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS053' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS052' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS051' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS050' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS049' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS048' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS047' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS046' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS045' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS044' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS043' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS042' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS041' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS038' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS037' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS031' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS029' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS028' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS027' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS026' : {'internal' : 11297.8, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS017' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS016' : {'internal' : 16083.33, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS015' : {'internal' : 22000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS014' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS013' : {'internal' : 12710, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS012' : {'internal' : 9000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS011' : {'internal' : 2680, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS010' : {'internal' : 41030, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS009' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS008' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'26TS007' : {'internal' : 30, 'ogures' : 31227.62, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 95462.38, 'prestation' : 0, 'ca_autre' : 0},
+'26TS002' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 33050, 'prestation' : 0, 'ca_autre' : 0},
+'26TS001' : {'internal' : 2479.52, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS379' : {'internal' : 19636.36, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS376' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 3680, 'prestation' : 0, 'ca_autre' : 0},
+'25TS374' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS373' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS372' : {'internal' : 32939.19, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS371' : {'internal' : 4850, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS370' : {'internal' : 6821.43, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS369' : {'internal' : -20000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS368' : {'internal' : 8100, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS367' : {'internal' : 13700, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS365' : {'internal' : 27200, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS364' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS363' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS362' : {'internal' : 19781.25, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS361' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS359' : {'internal' : 36464.91, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS358' : {'internal' : 47855.93, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS357' : {'internal' : 25971.05, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 16800, 'prestation' : 0, 'ca_autre' : 228.95},
+'25TS356' : {'internal' : 18920, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS355' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS354' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS353' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS352' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS351' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS350' : {'internal' : 15445.21, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS345' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS344' : {'internal' : 18640, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS343' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS340' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS339' : {'internal' : 40340, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS333' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS332' : {'internal' : 4135, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS331' : {'internal' : 10550, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS330' : {'internal' : 3.63797880709171E-12, 'ogures' : 50325.38, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 29674.62, 'prestation' : 0, 'ca_autre' : 0},
+'25TS329' : {'internal' : 24357.88, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 562.12},
+'25TS328' : {'internal' : 28120, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS327' : {'internal' : 37479.9, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 20.1},
+'25TS326' : {'internal' : 4315, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS325' : {'internal' : 8700, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS323' : {'internal' : 62850, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS322' : {'internal' : 0, 'ogures' : 65412.32, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 34587.68, 'prestation' : 0, 'ca_autre' : 0},
+'25TS321' : {'internal' : 47600, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS320' : {'internal' : 7791, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS319' : {'internal' : 14800, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS318' : {'internal' : 7164, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS317' : {'internal' : 18800, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS316' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS315' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS311' : {'internal' : 14551, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS307' : {'internal' : 7124.28, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 375.72},
+'25TS306' : {'internal' : 26425, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 975, 'prestation' : 0, 'ca_autre' : 0},
+'25TS303' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS300' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS299' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS298' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS295' : {'internal' : 3400, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS293' : {'internal' : 20856.58, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 93.42},
+'25TS292' : {'internal' : 20053, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS291' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS290' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS288' : {'internal' : 19450, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS287' : {'internal' : 24000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS284' : {'internal' : -312.5, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS283' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS282' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS280' : {'internal' : 18000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS277' : {'internal' : -39400, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS275' : {'internal' : -421.87, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS272' : {'internal' : 999.8, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS270' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS269' : {'internal' : 17500, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS268' : {'internal' : -790, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS267' : {'internal' : 13150, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS266' : {'internal' : 14530.55, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 69.45},
+'25TS264' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS256' : {'internal' : -2300, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS254' : {'internal' : 11675, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS253' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS251' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS244' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS240' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS238' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS228' : {'internal' : 56040, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS225' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS216' : {'internal' : 22000, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS206' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS204' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS196' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS195' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS174' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS170' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS161' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS142' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS141' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS137' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS136' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS134' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS131' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS130' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS126' : {'internal' : 2350, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS121' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS120' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS114' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS095' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS091' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS082' : {'internal' : 22800, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS016' : {'internal' : -813.56, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'25TS002' : {'internal' : 4518.95, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'24TS377' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'24TS357' : {'internal' : 25585, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 655},
+'24TS184' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+'23294' : {'internal' : 0, 'ogures' : 0, 'mc2i' : 0, 'inditto' : 0, 'autre_ST' : 0, 'prestation' : 0, 'ca_autre' : 0},
+        }
+
+        _logger.info("=========== DATA INITIALIZATION")
+
+        if self.rel_closing_date == datetime.date(2025, 12, 31):
+            _logger.info("========== Initialisation des cumuls au 31/12/2025 pour le projet %s" % self.rel_project_id.number)
+            if self.rel_project_id.number not in cumuls_31_12_2025.keys():
+                _logger.info("Ce projet n'est pas dans les données de Denis : %s" % self.rel_project_id.number)
+            else:
+                cumul_projet_31_12_2025 = cumuls_31_12_2025[self.rel_project_id.number]
+                _logger.info(str(cumul_projet_31_12_2025))
+                if self.type == 'internal_production' :
+                    self.progress_revenue_amount = cumul_projet_31_12_2025['internal'] + cumul_projet_31_12_2025['prestation']
+                elif self.type == 'other' :
+                    # il n'y a que le fournisseur FOURNISSEUR FRAIS DIVERS pour les lignes de type other pour décmebre 25 / janvier 26 et février 26 => donc attribution directe
+                    self.progress_revenue_amount = cumul_projet_31_12_2025['ca_autre']
+                    #rate = (self.progress_revenue_amount / self.target_project_revenue) if self.target_project_revenue else 0.0
+                    #self.outsourcing_product_qty = rate * self.target_project_outsourcing_product_qty
+                elif self.type == 'outsourcing' :
+                    p = self.env['project.progress'].search([('accounting_closing_id', '=', self.accounting_closing_id.id), ('type', '=', self.type)])
+                    if len(p) != 1 and (cumul_projet_31_12_2025['ogures']!=0 or cumul_projet_31_12_2025['mc2i']!=0 or cumul_projet_31_12_2025['inditto']!=0 or cumul_projet_31_12_2025['autre_ST']!=0) :
+                        _logger.info("Ce projet a plusieurs project.progress de type outsourcing : %s" % self.rel_project_id.number)
+                    else :
+                       self.progress_revenue_amount = cumul_projet_31_12_2025['ogures'] + cumul_projet_31_12_2025['mc2i'] + cumul_projet_31_12_2025['inditto'] + cumul_projet_31_12_2025['autre_ST']
+                       #rate = (self.progress_revenue_amount / self.target_project_revenue) if self.target_project_revenue else 0.0
+                       #self.outsourcing_product_qty = rate * self.target_project_outsourcing_product_qty
+                else : 
+                    _logger.info("Ce projet a un projet.progrss dont le type %s n'est pas pris en charge n'est pas géré : %s" % (self.type ,self.rel_project_id.numbe))
+                        
+
+        if self.rel_closing_date == datetime.date(2026, 1, 31):
+            _logger.info("========== Initialisation des CA périodique du mois de janvier 2026 pour le projet %s" % self.rel_project_id.number)
+            if self.rel_project_id.number not in ca_periode_01_2026.keys():
+                _logger.info("Ce projet n'est pas dans les données de Denis : %s" % self.rel_project_id.number)
+            else:
+                data_ca_projet = ca_periode_01_2026[self.rel_project_id.number]
+                _logger.info(str(data_ca_projet))
+                if self.type == 'internal_production' :
+                    #self.progress_revenue_amount_period = data_ca_projet['internal'] + data_ca_projet['prestation']
+                    self.progress_revenue_amount = self.rel_previous_progress_revenue_amount + data_ca_projet['internal'] + data_ca_projet['prestation']
+                elif self.type == 'other' :
+                    # il n'y a que le fournisseur FOURNISSEUR FRAIS DIVERS pour les lignes de type other pour décmebre 25 / janvier 26 et février 26 => donc attribution directe
+                    #self.progress_revenue_amount_period = data_ca_projet['ca_autre']
+                    self.progress_revenue_amount = self.rel_previous_progress_revenue_amount + data_ca_projet['ca_autre']
+                elif self.type == 'outsourcing' :
+                    p = self.env['project.progress'].search([('accounting_closing_id', '=', self.accounting_closing_id.id), ('type', '=', self.type)])
+                    if len(p) != 1 and (data_ca_projet['ogures']!=0 or data_ca_projet['mc2i']!=0 or data_ca_projet['inditto']!=0 or data_ca_projet['autre_ST']!= 0) :
+                        _logger.info("Ce projet a plusieurs project.progress de type outsourcing : %s" % self.rel_project_id.number)
+                    else :
+                       #self.progress_revenue_amount_period = data_ca_projet['ogures'] + data_ca_projet['mc2i'] + data_ca_projet['inditto'] + data_ca_projet['autre_ST']
+                       self.progress_revenue_amount = self.rel_previous_progress_revenue_amount + data_ca_projet['mc2i'] + data_ca_projet['inditto'] + data_ca_projet['autre_ST']
+                else : 
+                    _logger.info("Ce projet a un projet.progrss dont le type %s n'est pas pris en charge n'est pas géré : %s" % (self.type ,self.rel_project_id.numbe))
+                        
+        if self.rel_closing_date == datetime.date(2026, 2, 28):
+            _logger.info("========== Initialisation des CA périodique du mois de fevrier 2026 pour le projet %s" % self.rel_project_id.number)
+            if self.rel_project_id.number not in ca_periode_02_2026.keys():
+                _logger.info("Ce projet n'est pas dans les données de Denis : %s" % self.rel_project_id.number)
+            else:
+                data_ca_projet = ca_periode_02_2026[self.rel_project_id.number]
+                _logger.info(str(data_ca_projet))
+                if self.type == 'internal_production' :
+                    #self.progress_revenue_amount_period = data_ca_projet['internal'] + data_ca_projet['prestation']
+                    self.progress_revenue_amount = self.rel_previous_progress_revenue_amount + data_ca_projet['internal'] + data_ca_projet['prestation']
+                elif self.type == 'other' :
+                    # il n'y a que le fournisseur FOURNISSEUR FRAIS DIVERS pour les lignes de type other pour décmebre 25 / janvier 26 et février 26 => donc attribution directe
+                    #self.progress_revenue_amount_period = data_ca_projet['ca_autre']
+                    self.progress_revenue_amount = self.rel_previous_progress_revenue_amount + data_ca_projet['ca_autre']
+                elif self.type == 'outsourcing' :
+                    p = self.env['project.progress'].search([('accounting_closing_id', '=', self.accounting_closing_id.id), ('type', '=', self.type)])
+                    if len(p) != 1 and (data_ca_projet['ogures']!=0 or data_ca_projet['mc2i']!=0 or data_ca_projet['inditto']!=0 or data_ca_projet['autre_ST']!= 0) :
+                        _logger.info("Ce projet a plusieurs project.progress de type outsourcing : %s" % self.rel_project_id.number)
+                    else :
+                       #self.progress_revenue_amount_period = data_ca_projet['ogures'] + data_ca_projet['mc2i'] + data_ca_projet['inditto'] + data_ca_projet['autre_ST']
+                       self.progress_revenue_amount = self.rel_previous_progress_revenue_amount + data_ca_projet['mc2i'] + data_ca_projet['inditto'] + data_ca_projet['autre_ST']
+                else : 
+                    _logger.info("Ce projet a un projet.progrss dont le type %s n'est pas pris en charge n'est pas géré : %s" % (self.type ,self.rel_project_id.numbe))
+
